Реализовал: 1) Получение евента CommentKafkaEvent 2) Обновления комментариев в Post редиса 3) Реализовал отправление всех нужный сущностей в редис

+++ src/main/java/faang/school/postservice/model/redis/PostRedis.java
@@ -27,5 +27,5 @@ public class PostRedis {
     private LocalDateTime publishedAt;
     private LocalDateTime createdAt;
     private LocalDateTime updatedAt;
-    private int views;
+    private Long views;
 }
